refactor #170: use absolute imports instead of relative
