+++ app/stats/[slug]/page.tsx
@@ -6,7 +6,7 @@ import { ThemeProvider } from '@/hooks/useTheme'
 import { PlayersStats } from '@/interfaces/player'
 import { headers } from 'next/headers'
 import Error from '@/components/molecules/error'
-import TableWithControls from '@/components/tableWithControls'
+import TableWithControls from '@/components/organisms/tableWithControls'
 import ScoreboardSelector from '@/components/molecules/scoreboardSelector'
 import Info from '@/components/atoms/info'
 import Link from '@/components/atoms/link'
